Improve the path field type
The path class will now allow being instantiated using multiple path
parts, making it behave more like its parent pathlib.PurePath class.

The inference on whether to return a windows_path or posix_path class
instance is
now done based on the first path part it encounters that is an instance
of pathlib.PurePath.

For path parts that are a custom subclass of pathlib.PurePath (so not an
instance of PureWindowsPath or PurePosixPath), the presence of a '\' as
either separator or alternative separator will result in returning a
windows_path class instance.

(DIS-1977)